requirements.txt
praw
tavily-python
python-dotenv
firecrawl-py
langgraph
langchain-openai
azure-cognitiveservices-speech
pydub
